Fixed footer overlap
I fixed a footer overlap issue with the form on the edit profile page. I added an option on the frontend to edit email and username, backend still needs to be added. I removed a double footer on some pages. I added the tutorial link for the CRUD operations course to the main readme file.